requirements.txt
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==2.0.12
distlib==0.3.6
Django==3.2.16
django-domainname==0.1.1
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.4.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
pytz==2022.7
requests==2.27.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
zipp==3.6.0
